% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prettylog}
\alias{prettylog}
\title{prettylog}
\usage{
prettylog(prefix, ...)
}
\arguments{
\item{prefix}{A short label shown in brackets, e.g. a job name.}

\item{...}{Character strings passed to \code{paste0} to form the message.}
}
\description{
Emit a timestamped, prefixed log line to the console.
}
