% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr.R
\docType{data}
\name{qpcr}
\alias{qpcr}
\title{Auxiliary table for rq package, used in examples}
\format{
## `qpcr`
A tibble with 45 rows and 4 columns:
\describe{
  \item{Sample}{Sample names}
  \item{ITS2}{ITS2 locus Cq data}
  \item{pUC18}{pUC18 plasmid data}
  \item{DCq}{Delta Cq between ITS2 and pUC18}
}
}
\source{
artificial data
}
\usage{
qpcr
}
\description{
Auxiliary table for rq package, used in examples
}
\keyword{datasets}
