% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_rq.r
\name{rq}
\alias{rq}
\title{Calculates the Relative Quantity of a qPCR set}
\usage{
rq(data, normalizer, reference, treatment, delimiter = "-")
}
\arguments{
\item{data}{A tibble}

\item{normalizer}{The name of the normalizer samples.}

\item{reference}{The name of the reference target.}

\item{treatment}{The name of the treatment target.}

\item{delimiter}{The delimiter character separating sample name and replica. Defaults to "-".}
}
\value{
A tibble
}
\description{
It takes a tibble of five columns: Sample, Target Name, Normalizer Name
and DCq (Delta Cq), to calculate the relative quantity and return another
tibble containing such data. Take a look at qpcr data available with this
package for and example.
}
\examples{
rq(qpcr, "pUC18", "WT", "ITS2")
}
