% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowPrinter.R
\name{WorkflowPrinter}
\alias{WorkflowPrinter}
\title{WorkflowPrinter}
\description{
Renders the job DAG as an ASCII tree for inspection.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WorkflowPrinter-print_workflow}{\code{WorkflowPrinter$print_workflow()}}
\item \href{#method-WorkflowPrinter-clone}{\code{WorkflowPrinter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowPrinter-print_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowPrinter-print_workflow}{}}}
\subsection{Method \code{print_workflow()}}{
Print the DAG rooted at all jobs with no upstream dependencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowPrinter$print_workflow(orchestrator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orchestrator}}{A JobOrchestrator instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowPrinter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowPrinter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowPrinter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
